Finish new style DriverComponent and adjust LS350Driver
The new style DriverComponent allows for more flexibility in processing
commands. In the old style driver there was no way for the commands to
store information on the DriverComponent as the commands were in the Driver
class.

The new DriverComponent now contains all the commands and the commands
have the DriverComponent passed to them so that they can modify and store
information on the DriverComponent object.