Fix execution of built up queue events
If a component communicates with the server multiple times it can leave
many events in the queue. If we need to restart the component it will
have to process the events in the queue before processing new events.
This would lead to old events being generated.

Now we clear the queue by deleting it before restarting it in the
component. This leaves an empty queue with which to communicate.